==> 04_data_prep.ipynb <==
# Data prep for analyses  

#### Notebook objectives:  
- Filter, clean and transform the raw AskDocs subreddit data
- Save the resulting analysis dataset  

#### Steps:  
First, we need to work out the needed data prep steps on a smaller data subset. We'll use the 2017 raw dataset for that.  
1. [Load 2017 data](#Load-2017-data)  
2. [Filter out irrelevant messages](#Filter-out-irrelevant-messages)  
3. [Consolidate post and cross-post content](#Consolidate-post-and-cross-post-content)
   - [Get cross-post subreddit names](#Get-cross-post-subreddit-names)  
4. [Consolidate the title and post body as the full user question](#Consolidate-the-title-and-post-body-as-the-full-user-question)
5. [Fix timestamp formats](#Fix-timestamp-formats)
6. [Update the selected fields list](#Update-the-selected-fields-list)

Next, we apply the above data prep steps to the entire dataset.  
7. [Load all data](#Load-all-data)  
8. [Apply data prep steps](#Apply-data-prep-steps)  
9. [Save the analysis dataset](#Save-the-analysis-dataset)
   

import pickle
import pandas as pd
import numpy as np
from IPython.display import display, HTML, Markdown, clear_output
import ipywidgets as widgets
import time 

DATA_PATH = 'data/'
OUTPUT_PATH = 'output/'

## Load 2017 data

with open(DATA_PATH + "reddit_askdocs_submissions_2017.pkl", "rb") as f:
    d_2017 = pickle.load(f)

df_2017 = pd.DataFrame(d_2017)

df_2017.head()

df_2017['author_flair_text'].value_counts(dropna=False, normalize=True)

## Filter out irrelevant messages  

Filtering out submissions not relevant to this analysis, like mod posts and banned posts. We want to analyze submissions that are medical questions from users. Mod posts are typically announcements, and bannded posts were deemed unrelated or inappropriate by subreddit moderators, so those types of posts can be filtered out as not relevant to analysis.

fname = 'reddit_submissions_selected_fields.csv'
selected_fields_df = pd.read_csv(fname)

display(Markdown(selected_fields_df[selected_fields_df['analysis_role'] == 'f'].to_markdown()))

filter_fields = selected_fields_df[selected_fields_df['analysis_role'] == 'f']['field_name'].to_list()
filter_fields

def filter_submissions(df):
    original_rows_count = len(df)
    print(f'Original rows count: {original_rows_count}')
    
    filters = (df['stickied'] == True) | (df['distinguished'] == 'moderator') \
        | (df['banned_by'] == 'moderators')
    df = df.drop(df[filters].index)
    
    print(f'Filtered rows count: {len(df)}')
    print(f'% of rows filtered out: {(1-len(df)/original_rows_count)*100:.2f}')
    
    return df

df_2017 = filter_submissions(df_2017)

## Consolidate post and cross-post content

display(Markdown(selected_fields_df[selected_fields_df['analysis_role'] == 't']
                 .to_markdown()))

First, we need to clean up the `selftext` post body field a little.

df_2017['selftext'] = df_2017['selftext'].str.strip() 

df_2017['selftext'].str[:40].value_counts(dropna=False).head(20)

def clean_selftext_blanks(txt):
    if type(txt) == str:
        r = '' if txt in ['', '[deleted]', '[removed]'] else txt
    elif np.isnan(txt):
        r = ''
    else:
        r = txt
    return r

df_2017['selftext'] = df_2017['selftext'].apply(clean_selftext_blanks)

df_2017['selftext'].str[:40].value_counts(dropna=False).head(10)

df_2017[df_2017['selftext'] == '']['crosspost_parent'].isna()\
    .value_counts(dropna=False).head(20)

df_2017[df_2017['selftext'] != '']['crosspost_parent'].isna()\
    .value_counts(dropna=False).head(20)

df_2017[df_2017['selftext'] == '']['crosspost_parent_list'].isna()\
    .value_counts(dropna=False).head(20)

df_2017[df_2017['selftext'] != '']['crosspost_parent_list'].isna()\
    .value_counts(dropna=False).head(20)

df_2017['crosspost_parent_list'].isna().value_counts(dropna=False)

df_2017[df_2017['crosspost_parent_list'].notna()]['selftext'].str[:40]\
    .value_counts(dropna=False)

After cleaning up the `selftext` field, the value counts above show that the crosspost content is available only when the cleaned `selftext` field is blank.

Now we can go ahead and use the crosspost content to fill in the blank post body content in the `selftext` field, if the crosspost content is available.

We can get the submission body and the crosspost subreddit from the dict inside the `crosspost_parent_list` field, all other relevant fields in the dict are duplicates of the submission-level fields. 

def get_crosspost_body(lst):
    # We'll take the body text of the first non-blank/non-deleted cross-post
    r = ''
    if type(lst) == list:
        for i in lst:
            txt = i['selftext'].strip()
            if txt.lower() not in ['', '[deleted]', '[removed]']:
                r = txt
                break
    return r

crosspost_fix_condition = (
    (df_2017['selftext'] == '') 
    & (df_2017['crosspost_parent_list'].notna())
)

df_2017['selftext'] = np.where(
    crosspost_fix_condition, 
    df_2017['crosspost_parent_list'].apply(get_crosspost_body), 
    df_2017['selftext']
)

df_2017[df_2017['crosspost_parent_list'].notna()]['selftext']

### Get cross-post subreddit names  
The names of the cross-post subreddits could provide interesting info, so let's pull them out here into a variable as well.

def get_crosspost_subreddits(lst):
    r = ''
    if type(lst) == list:
        subreddits_list = []
        for i in lst:
            sr = i['subreddit']
            subreddits_list.append(sr)
            
        r = ', '.join(subreddits_list)
    
    return r

df_2017['crosspost_subreddits'] = df_2017['crosspost_parent_list']\
    .apply(get_crosspost_subreddits)

df_2017['crosspost_subreddits'].value_counts(dropna=False)

## Consolidate the title and post body as the full user question  
Users sometimes ask all or part of their question in the title while the post body may be blank or may continue from the title.  
To get the full user question, we will combine the `title` and `selftext` fields into a single text string.  

df_2017['selftext'].str[:40].value_counts(dropna=False)

df_2017[df_2017['selftext'] == '']['title'].str[:40].value_counts(dropna=False)

To make it easier to visually separate the title and post body in the combined text field, we'll use a pipe (`|`) separator.

df_2017['full_post_text'] = (
    df_2017['title'].str.strip() 
    + ' | '
    + df_2017['selftext'].str.strip()
)

df_2017['full_post_text'].str[:40].value_counts(dropna=False).head(20)

(df_2017['full_post_text'] == '').value_counts()

## Fix timestamp formats  

timestamp_fields = ['created_utc', 'edited']

df_2017['created_utc'].head()

pd.to_datetime(df_2017['created_utc']).head() # the default in pandas is 'ns'

pd.to_datetime(df_2017['created_utc'], unit='s').head()

pd.to_datetime(df_2017[df_2017['edited'].notna()]['edited']).head()

pd.to_datetime(df_2017[df_2017['edited'].notna()]['edited'], unit='s').head()

The raw timestamp fields are in epochs with seconds as units. The default units in pandas is 'ns'. The two fields appear to be in the same format, but the `edited` field is named less informatively.  
To make these timestamp fields easier to work with in pandas, let's do the following:  
- Convert them to 'ns' units, since that's the default unit in pandas. This way, we won't have to remember to explicitly specify the units each time we cast these fields to datetime in pandas.  
- Rename both fields to make it more clear that they are UTC timestamps stored in nanosecond epochs.  

df_2017['created_utc_ns_dt'] = df_2017['created_utc'] * 1_000_000_000

df_2017['edited_utc_ns_dt'] = df_2017['edited'] * 1_000_000_000

pd.to_datetime(df_2017['created_utc_ns_dt']).head()

pd.to_datetime(df_2017['edited_utc_ns_dt']).value_counts(dropna=False)

## Update the selected fields list  
Let's add the new fields we created for analyses and drop any fields we no longer need.

display(Markdown(selected_fields_df[~selected_fields_df['analysis_role'].isin(['f', 't'])].to_markdown()))

list(selected_fields_df[~selected_fields_df['analysis_role'].isin(['f', 't'])]['field_name'])

# Final fields selection

selected_fields = [
    'author',
    'author_flair_text',
#    'created_utc', # replaced with the new created_utc_ns_dt field below
    'domain',
    'full_link',
    'id',
    'locked',
    'num_comments',
    'num_crossposts',
    'over_18',
#    'pinned',
    'score',
    'selftext', # edited to include the body of crossposted posts where needed
#    'spoiler',
    'title',
    'url',
#    'edited', # replaced with the edited_utc_ns_dt field below
#    'author_fullname',

    
    # newly added fields
    'crosspost_subreddits',
    'full_post_text',
    'created_utc_ns_dt',
    'edited_utc_ns_dt',
]

df_2017[selected_fields].head()

## Load all data  

with open(DATA_PATH + "reddit_askdocs_submissions_2018_to_20220121.pkl", "rb") as f:
    d_ge2018 = pickle.load(f)

df = pd.concat(
    [
        df_2017, 
        pd.DataFrame(d_ge2018[2018]),
        pd.DataFrame(d_ge2018[2019]),
        pd.DataFrame(d_ge2018[2020]),
        pd.DataFrame(d_ge2018[2021]),
        pd.DataFrame(d_ge2018[2022])
    ]
)

del d_2017, df_2017, d_ge2018

df.head()

## Apply data prep steps

1. Filter out irrelevant messages

df = filter_submissions(df)

2. Consolidate post and cross-post content
   

df['selftext'] = df['selftext'].str.strip() 

df['selftext'] = df['selftext'].apply(clean_selftext_blanks)

crosspost_fix_condition = (
    (df['selftext'] == '') 
    & (df['crosspost_parent_list'].notna())
)

df['selftext'] = np.where(
    crosspost_fix_condition, 
    df['crosspost_parent_list'].apply(get_crosspost_body), 
    df['selftext']
)

4. Get cross-post subreddit names

df['crosspost_subreddits'] = df['crosspost_parent_list']\
    .apply(get_crosspost_subreddits)

5. Consolidate the title and post body as the full user question

df['full_post_text'] = (
    df['title'].str.strip() 
    + ' | '
    + df['selftext'].str.strip()
)

# check that we don't have blanks in full_post_text
(df['full_post_text'] == '').value_counts()

6. Fix timestamp formats

df['created_utc_ns_dt'] = df['created_utc'] * 1_000_000_000

df['edited_utc_ns_dt'] = df['edited'] * 1_000_000_000

7. Filter to the final selected fields

df = df[selected_fields]

df.info()

## Save the analysis dataset

# Save file with on-the-fly compression
df.to_csv(
    DATA_PATH + 'reddit_askdocs_submissions_2017_to_20220121_analysis_ds.zip', 
    index=False
)
